refactor(settings): one primitive for self-anchoring controls (#573)
A catalog control that anchors itself — no header/card split to
place the reveal pair across — had no primitive to reach for, so
`GapRow` and `BarEditorPicker` each hand-encoded the placement,
and the confinement guard's allow-list was on its way to a third
copy of the rule it exists to protect.

Add `.settingsAnchor(_ control:)`, which applies both halves from
one `SettingsControl`. Taking the descriptor rather than an id is
what closes the two silent failures a hand-written pair allows: a
literal id (a scroll target no search result points at) and a
forgotten half (a flash with nowhere to scroll, or the reverse).
Neither crashes — the hit simply lands nowhere.

Both existing bespoke controls move onto it rather than wait for
a third, so the allow-list SHRINKS to the shapes that genuinely
must split the pair across chrome. `BarEditorPicker` splits at
the Button boundary today; with `.plain` the button's frame is
its label's frame, so the wash lands in the same rect from
outside. `SettingsSection` loses its `anchorID ?? ""` flash
sentinel — an id no reveal can ever name, reading as a second,
disagreeing half.

Guards for what the type cannot reach: per blessed file the two
argument SETS must be equal, non-empty and literal-free (sets,
not counts — `SettingsDisclosure` anchors in both chrome branches
and flashes once), and the primitive must stay reached-for, or
the confinement guard becomes a dead end. All four failure modes
verified red by injection.

Splits, both files one line under the ceiling: the catalog
declarations divide on the seam the sidebar already draws
(`thisProfile` / `wholeApp`), and the site guards split by
direction. `isCatalogFile` becomes a `SettingsCatalog*` prefix
test so a further slice needs no allow-listing. The call-shape
walker moves to `SourceScan+Arguments` on the standing drift bar
— a second consumer appeared, and a copy that normalizes
differently reads a different string from the same source and
passes for the wrong reason.

fixes #573
