+++ Sources/KiwiDesk/Settings/Components/Appearance/GapsEditor.swift
@@ -211,17 +211,16 @@ struct GapsEditor: View {
 /// catalog descriptor, so the row is a search anchor by existing
 /// — the first drawer-interior controls the catalog reaches
 /// (#277); a reveal expands the drawer (`SettingsDisclosure`)
-/// and lands here. The wash covers the whole row: unlike a
-/// section card there is no expanded content below the label to
-/// tint by accident.
+/// and lands here. Self-anchoring, so it takes `.settingsAnchor`
+/// whole: the wash covers the same rect as the scroll target
+/// because, unlike a section card, there is no expanded content
+/// below the label to tint by accident.
 private struct GapRow: View {
     let control: SettingsControl
     @Binding var value: CGFloat
 
     var body: some View {
-        row
-            .searchFlash(control.id)
-            .searchAnchor(control.id)
+        row.settingsAnchor(control)
     }
 
     private var row: some View {
